python/tracing/pydantic-ai/requirements.txt
pydantic-ai>=0.2.0
respan-ai
respan-instrumentation-pydantic-ai
python-dotenv
